extract webpack common config file
